pwkit/msmt.py: import modernized Lval and Textual code
This is a meld of a lot of work going into implementing Lvals and Textual with
a unified math system. It took me many tries to get the Lval behavior right
and to figure out how to provide a unified math library without a lot of
redundant implementations lying around. I believe that this version gets
things pretty much as right as I can manage.