pwkit/cli/wrapout.py: redirect subprocess input to /dev/null
I don't think it makes sense to use wrapout for interactive programs, so it's
best to make stdin /dev/null in case some part of the child tries to read
input. This also helps because weird stuff can happen if the child is hooked
up to a terminal. However, that means that if we get a SIGINT we need to
forward it on to the child ourselves. We weren't really handling these right
before anyway. (... I think ... I suck at TTYs and signals.)