pwkit/parallel.py: add a new ppmap() parallel function
This is a "partially Pickling" map function that provides a mechanism for
working around certain limitations in the multiprocessing module, namely that
way that all of the items involved a Pool.map() call needed to be Pickle-able,
including the mapping function itself. ppmap provides an alternate mechanism
as documented in some detail in the code in the commit.