pwkit/cli: clever (?) system for asking a program to print a traceback
Since I learned that Python signal handlers are passed the Python stack when
they're called, I had the idea that it might be useful to provide a simple
hook so that one can print the stack as a debugging feature. The new
pwkit.cli.backtrace_on_usr1() does that. I haven't seen this idea elsewhere,
nor have I tested it much, so maybe it's not as bright as I think. But there
are definitely times when I have a long-running program and I either want to
check that it's doing what I think it should be, or I realize it's stuck and
want to figure out where.