docs: initial restructure of Foundations section
I realized that there's no reason that the doc structure has to track the
module structure (duh). The starting point is 'one doc page for each module',
but the intent is to morph that into having doc pages that make sense as
prose, and rearranging the API documentation to fit in wherever it belongs.