pwkit/environments/sas/data.py: do not convert counts to an integer
If the light curve is given explicit time bounds, there may be bins in which
`rate` is NaN. Trying to compute `counts` from NaN and converting to integer
causes an exception to be raised. We could use -1 or something as a mask
value, but it seems better to preserve the NaN-ness -- since we're
constructing the integer from a float anyway, I don't think using integers
actually helps us with precision. It's nice for the counts to be all integer-y
but it's also nice to be able to express "no data" cleanly.