pwkit/cli/imtool.py: fix pop_option usage
I don't need to add a dummy argv[0] entry because pop_option is cavalier about
that anyway. This un-breaks the -p option to fitsrc, because pop_option with a
temporary array wouldn't affect `args` and so the check for the correct number
of arguments would fail. I should probably stop rolling my own arg-handling
code.