pwkit/environments/jupyter: initial work on Jupyter ergonomics
The main thing here is a more sensible command-line interface that just
launches a server in the background in a standardized way. It doesn't log
stuff to your terminal, it doesn't need backgrounding, it always starts in
$HOME for consistency, and it only starts a new server if needed. Other
helpers can auto-kill the server too. I think this is the first step to me
being more eager to use the notebook interface instead of console IPython.