docs/source/foundations/parallel.rst: write OK docs
Aha, I think I've figured out how to handle things. I wanted to use automodule
for my module docs, but it seemed like interleaving that plus explanatory text
was going to be annoying and require a lot of indentation. But the
`currentmodule` directive seems to do what I want. The `autosummary` directive
is also helpful although I need to tweak its rendering to not make tables that
scroll off the right side of the screen.