pwkit/cli/latexdriver.py: have 'bib' ignore missing entries in '-E' mode
If we're using the merged-BibTeX mode with bibtools, it's reasonable to expect
that the document will end up having references to papers unknown to the
user's bibtools library. Therefore, let's default to telling bibtools to not
raise an error for missing references when in this mode.

Requires bibtools commit 51d35c4 (2016 Feb 22) or later.