pwkit/environments/casa/spwglue.py: add corr_to_main option
This does the glue while renaming CORRECTED_DATA to DATA, effectively applying any calibrations
that may have been performed on the input dataset. I want this for my new approach to applying
VLA switched-power calibrations, which are per-spw in a way where I don't want to just average
over them. (And I don't want to have to figure out the code to do that.)