Get "autoattribute" working in Sphinx.
I ran into a roadblock trying to get Sphinx's autodoc extension to work with
class attributes. I eventually discovered that the problem was Sphinx bug 1641
(https://github.com/sphinx-doc/sphinx/issues/1641). I use the print_function
future in all of my code. But to document class attributes, Sphinx opens up
the source code files and parses them, and the module it uses to do this
parsing is not away of the print_function future. It freaks out if it sees
print used with a construct like `file=f`, so Sphinx gives up on parsing, and
is thus unable to see any per-attribute docstrings.

The gross solution is to give the print function another name to prevent the
parser from freaking out. I don't aim to print too many things in library code
so hopefully this won't be too annoying.