pwkit/environments/ciao/data.py: improve event plotting
We add a mode where you can bin up individual events by the median energy, with
an estimate of the error of the median. This is useful for looking at all of
the events on a CCD, which can show you spikes in the background rate but has
too many events to plot individually. Doing this nicely requires the GTI
intersection bit that I just added in bblocks.