pwkit/environments/casa: clarify that dft* only work on whole spectral windows
We let users use sub-spw selections like `spw=0:5~10`, but they were being
silently ignored. That's bad. We still ignore them, but are not silent about
it anymore. Would be nice to figure out how to properly process sub-spw
selections properly but at the moment I don't know how to do it, and it's
never been that urgent of a feature for me.