docs: get autoattribute working on dftdynspec
Couldn't let this one go.

By manually running `sphinx-build` with the `-vvv` option, I discovered that
Sphinx was indeed failing to parse the `dftdynspec.py`, which was why it
wasn't able to pull out docstrings. It seems pretty lame that this is a silent
failure.

The problem seems to be with lines of the form `from ...foo import bar`. I
wouldn't be shocked if there was some confusion with the ellipsis token or
something. By adding a space between two dots, both Python itself and Sphinx's
tokenizer are happy.