pwkit/latex.py: Python 3 unicode/bytes fixes, sort of
I am torn here. More and more I am preferring the idea of writing TeX code out
as Unicode strings. But, if we're not using a Unicode-aware engine, we need a
way to translate Unicode characters into their non-Unicode TeX
representations, and the text/bytes distinction is one of the few tools we
have available in Python to try to enforce correct escaping. Almost makes you
long for a more strongly-typed language ...