pwkit/data_gui_helpers.py: fix for MaskedArray semantics in newer Numpy
It seems that when we use a function like np.subtract() with a MaskedArray,
the mask now gets converted to a scalar, presumably an optimization if the
mask is all False or whatever. This breaks the code that we had before. (Note
this happens even if using np.ma.subtract().) If we operate on the `.data`
portion of the MaskedArray, things work.