pwkit/environments: yet more bytes/unicode fixing
The `os.environ` data structure is yet another thing that is bytes in Python 2
and Unicode in Python 3. This doesn't interact well when we use the
`unicode_literals` future mode, so get rid of it, and try in CASA to be better
about passing `str`s to the environment functions as opposed to things that
are fixed to be either bytes or unicode in particular. Apparently solve's
Tarraneh's problem.d