Work on de-bytes-ifying the LaTeX code.
The various bits of pwkit's LaTeX code tried to impose a system in which
arbitrary input Unicode text was represented as Python Unicode data (`unicode`
in Python 2), whereas LaTeX-escaped text ready to hand to the LaTeX engine was
represented as bytes (`str` in Python 2, `bytes` in Python 3). Having two
different types for unescaped vs. escaped text was appealing, but really gets
to be a hassle in Python 3. Also, what we're outputting really is text, and
nowadays I exclusively use Tectonic or XeTeX, which accept Unicode inputs
anyway.

So in this commit we start excising the use of bytes types in `pwkit.latex`
and related functions. I expect that various corner cases are broken, and I am
very worried about "double-escaping" input text (so that "\foo" becomes
"\\\\foo" by the time it hits the output file). If this were Rust, I'd create
a shim type `struct LatexEncodedString(String)` or something, but ... this is
not Rust.

What I've tried to do is make it so that if you pass in ASCII bytes to a
function that normally does a LaTeX escaping pass, it will just pass the bytes
through, decoding the ASCII into a Python Unicode string object. This will
hopefully give a useful mechanism for callers to explicitly assert that have
already thought about LaTeX escaping.

Overall I expect that doing things this way will create a steady stream of
bugs that need squashing, but that sticking with LaTex-as-bytes would create
an even steadier stream of bugs.