pwkit/environments/casa/tasks.py: fix bpplot for varying nchan
The bpplot algorithm assumed that every spw had the same number of channel,
which isn't guaranteed. Fix it to not make this assumption. This makes it so
that we have to load CPARAM and FLAG cell-by-cell, instead of all in one go,
but I think that's just how things have to be.